+++ __tests__/src/test-utils/configs.ts
@@ -59,7 +59,7 @@ export const getConfigs = (port: number) => {
   };
 
   return {
-    handlerConnection: handlerConnection,
+    handlerConnection,
     outboxConfig,
     inboxConfig,
   };